Extract FeedLoaderStub int a shared scope to remove duplication
